tests/wpt/tests/import-maps/acquiring/script-tag.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const log = [];
</script>
<script type="module" src="../resources/empty.js?pipe=trickle(d1)"></script>
<script type="importmap">
{
  "imports": {
    "../resources/log.js?pipe=sub&name=A": "../resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
promise_test(() => {
  return import("../resources/log.js?pipe=sub&name=A")
    .then(() => assert_array_equals(log, ["log:B"]))
  },
  'After <script type="module"> import maps work fine');
</script>
